moved and renamed logout-feedback file
